chore: remove unused schemas, files
